rework blacklisting and backoff
This commit is to make sure we never blacklist ourselves into a 503
situation and to make backoff immediate but less aggressive. Whereas
before it took 10 failures before any backoff occurred now we immediately
back off on the first failure. There are a few reasons, including that
10 is a pretty arbitrary cutoff when load may be different in different
environments. Exponential backoff is also too fast for transient
problems because there may be multiple outstanding requests to the host
when it fails, which would cause quite a large backoff immediately.

 * Keep track of how many failed peers we have.
 * Don't skip hosts if everything is marked down, just try and see.
 * Use a power function for backoff with no threshold.
   This should be quicker to backoff and quicker to recover.
 * Fix up the peer selection which should be a loop
 * Make #include of time.h explicit and add math.h
 * Link with -lm for math pow() function